Add navbar component to Todo app
