+++ src/App.jsx
@@ -1,9 +1,8 @@
 
 import { useEffect, useState } from 'react'
 import './App.css'
-import { TodoForm, TodoItem, ThemeBtn } from './components/index.js'
-import { TodoProvider, ThemeProvider } from './contexts/index.js'
-import todoIcon from "/todoIcon.png"
+import { TodoForm, TodoItem, Navbar } from './components/index.js'
+import { TodoProvider } from './contexts/index.js'
 function App() {
 
   const [todos, setTodos] = useState([])
@@ -41,29 +40,13 @@ function App() {
   }, [todos])
 
 
-  const [themeMode, setThemeMode] = useState("light")
-
-  const lightTheme = () => {
-    setThemeMode("light")
-  }
-  const darkTheme = () => {
-    setThemeMode("dark")
-  }
-
-  //actual change in theme using JS
-  useEffect(() => {
-    document.querySelector('html').classList.remove("light", "dark")
-    document.querySelector('html').classList.add(themeMode)
-  }, [themeMode])
-  
 
 
   return (
     <TodoProvider value={{ todos, addTodo, updateTodo, deleteTodo, toggleCompleted }}>
-
+      <Navbar />
       <div className="bg-[#ffffff] dark:bg-[#202124] min-h-screen py-8">
         <div className="w-full max-w-2xl mx-auto  rounded-lg px-4 py-3  text-[#212121] dark:text-[#fff]">
-          <h1 className="text-3xl font-bold text-center mb-8 mt-2 flex items-center justify-center">Manage Your Todos <img src={todoIcon} alt="Todo List Icon" className='w-[3rem] ml-2' /></h1>
           <div className="mb-8">
             <TodoForm />
           </div>
@@ -75,11 +58,6 @@ function App() {
               </div>
             ))}
           </div>
-          <ThemeProvider value={{themeMode, lightTheme, darkTheme}}>
-            <div className='fixed right-4 bottom-4 flex p-2'>
-              <ThemeBtn />
-            </div>
-          </ThemeProvider>
         </div>
       </div>
     </TodoProvider>
