+++ src/components/index.js
@@ -1,5 +1,6 @@
 import TodoForm from "./TodoForm";
 import TodoItem from "./TodoItem";
 import ThemeBtn from "./ThemeBtn";
+import Navbar from "./Navbar";
 
-export {TodoForm,TodoItem, ThemeBtn}
+export { TodoForm, TodoItem, ThemeBtn, Navbar }
